master to dana

+++ src/main/java/com/example/Instagram.java
@@ -19,15 +19,17 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.CrossOrigin;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import com.example.model.DataEntry;
+import com.example.model.Flag;
+import com.example.others.FlagStatus;
 import com.example.repo.ClientRepository;
 import com.example.repo.DataEntryRepository;
+import com.example.repo.ErrorLogsRepository;
 import com.example.repo.FlagRepository;
 import com.example.repo.IntervalRepository;
 import com.example.repo.LastEntryRepository;
@@ -50,6 +52,8 @@ public class Instagram {
 	DataEntryRepository dataRepo;
 	@Autowired
 	IntervalRepository intervalRepo;
+	@Autowired
+	ErrorLogsRepository errorRepo;
 
 	Entity entity = new Entity();
 	String RETURNURL = "";
@@ -61,18 +65,18 @@ public class Instagram {
 		return "preadmin";
 	}
 
-	/* HANDLE POST REQUEST FROM VIEW, see preadmin.html */
-	@PostMapping(value = "/admin", consumes = { MediaType.APPLICATION_FORM_URLENCODED_VALUE,
-			MediaType.APPLICATION_JSON_UTF8_VALUE }, produces = { MediaType.APPLICATION_ATOM_XML_VALUE,
-					MediaType.APPLICATION_JSON_UTF8_VALUE })
-	public String newIndex(@RequestParam("appId") String appId, @RequestParam("appSecret") String appSecret,
-			Model model) {
-		
-		System.out.println(" ===== ADMIN CALLED ======");
-		model.addAttribute("appId", appId);
-		model.addAttribute("appSecret", appSecret);
-		return "admin";
-	}
+	/* HANDLE POST REQUEST FROM VIEW, see preadmin.html --- NOT USED ANYMORE */
+	/*
+	 * @PostMapping(value = "/admin", consumes = {
+	 * MediaType.APPLICATION_FORM_URLENCODED_VALUE,
+	 * MediaType.APPLICATION_JSON_UTF8_VALUE }, produces = {
+	 * MediaType.APPLICATION_ATOM_XML_VALUE, MediaType.APPLICATION_JSON_UTF8_VALUE
+	 * }) public String newIndex(@RequestParam("appId") String
+	 * appId, @RequestParam("appSecret") String appSecret, Model model) {
+	 * 
+	 * System.out.println(" ===== ADMIN CALLED ======"); model.addAttribute("appId",
+	 * appId); model.addAttribute("appSecret", appSecret); return "admin"; }
+	 */
 
 	/*
 	 * @RequestMapping(method = RequestMethod.POST, consumes = {
@@ -90,8 +94,10 @@ public class Instagram {
 	String indexPost(@RequestParam Map<String, String> paramMap, Model model) {
 		RETURNURL = paramMap.get("return_url");
 
-		/*model.addAttribute("appId", entity.APP_ID);
-		model.addAttribute("appSecret", entity.APP_SECRET);*/
+		/*
+		 * model.addAttribute("appId", entity.APP_ID); model.addAttribute("appSecret",
+		 * entity.APP_SECRET);
+		 */
 		model.addAttribute("callbackUrl", entity.CALLBACKURL);
 		return "preadmin";
 	}
@@ -113,12 +119,12 @@ public class Instagram {
 		ArrayList<HashMap<String, String>> hashList = new ArrayList<>();
 		try {
 
-			JSONObject output = calling.hit(entity.getAccTokenApi(appId, appSecret) + token, "GET");
+			JSONObject output = calling.hitAuth(entity.getAccTokenApi(appId, appSecret) + token, "GET", errorRepo, appId +  " - Submit");
 			accToken = output.getString("access_token");
 
 			try {
 
-				JSONObject outputAcc = calling.hit(entity.GET_ACC_ID_API + accToken, "GET");
+				JSONObject outputAcc = calling.hitAuth(entity.GET_ACC_ID_API + accToken, "GET", errorRepo, appId +  " - Submit");
 				JSONArray igData = outputAcc.getJSONArray("data");
 				if (outputAcc != null) {
 					for (int i = 0; i < igData.length(); i++) {
@@ -199,51 +205,70 @@ public class Instagram {
 		// String option = "1";
 
 		List<DataEntry> dataEntry = dataRepo.findByCifAccountId(accountId);
+
+		Flag flagging = flagRepo.findByCifAccountId(accountId);
+		
 		List<DataEntry> willbeDelete = new ArrayList<>();
 		boolean alreadyFull = false;
 		int extCounter = 0;
 
 		System.out.println(dataRepo.count());
+		
+		HttpStatus responseCode;
+		
+		if (flagging == null) {
+			flagging = newAccountFlag(accountId);
+		}
+		
+		if (flagging.getCifStatus().equals(FlagStatus.REAUTH.toString())) {
+			responseCode = HttpStatus.UNAUTHORIZED;
+		} else {
+			for (int i = 0; i < dataEntry.size(); i++) {
 
-		for (int i = 0; i < dataEntry.size(); i++) {
-
-			@SuppressWarnings("unchecked")
-			ArrayList<Object> cifJsonData = gson.fromJson(dataEntry.get(i).getCifJsonData(), ArrayList.class);
-			if (!alreadyFull) {
-				for (int j = 0; j < cifJsonData.size(); j++) {
-					if (extCounter >= 199) {
-						alreadyFull = true;
-						extResourceRest.add(cifJsonData.get(j));
+				@SuppressWarnings("unchecked")
+				ArrayList<Object> cifJsonData = gson.fromJson(dataEntry.get(i).getCifJsonData(), ArrayList.class);
+				if (!alreadyFull) {
+					for (int j = 0; j < cifJsonData.size(); j++) {
+						if (extCounter >= 199) {
+							alreadyFull = true;
+							extResourceRest.add(cifJsonData.get(j));
+						} else {
+							extResource.add(cifJsonData.get(j));
+							extCounter++;
+						}
+					}
+					if (extResourceRest.size() > 0) {
+						System.out.println("===== UPDATE DB WITH ID: " + dataEntry.get(i).getId() + " =====");
+						doSaveDataEntryDb(dataEntry.get(i).getId(), dataEntry.get(i).getCifAccountId(),
+								dataEntry.get(i).getCifPostId(), extResourceRest);
+						extResourceRest = new ArrayList<>();
 					} else {
-						extResource.add(cifJsonData.get(j));
-						extCounter++;
+						willbeDelete.add(dataEntry.get(i));
 					}
 				}
-				if (extResourceRest.size() > 0) {
-					System.out.println("===== UPDATE DB WITH ID: " + dataEntry.get(i).getId() + " =====");
-					doSaveDataEntryDb(dataEntry.get(i).getId(), dataEntry.get(i).getCifAccountId(),
-							dataEntry.get(i).getCifPostId(), extResourceRest);
-					extResourceRest = new ArrayList<>();
-				} else {
-					willbeDelete.add(dataEntry.get(i));
-				}
 			}
-		}
-		response.put("external_resources", extResource);
+			response.put("external_resources", extResource);
 
-		for (int i = 0; i < willbeDelete.size(); i++) {
-			dataRepo.delete(willbeDelete.get(i));
+			for (int i = 0; i < willbeDelete.size(); i++) {
+				dataRepo.delete(willbeDelete.get(i));
+			}
+
+			if (dataRepo.count() <= 2) {
+				ThreadingTicket ticketThread = new ThreadingTicket(accountId, token, option, flagRepo, lastRepo, dataRepo,
+						intervalRepo, errorRepo);
+				ticketThread.start();
+			} else {
+				System.out.println("===== Still too many rows at DB =====");
+			}
+			responseCode = HttpStatus.OK;
 		}
 
-		if (dataRepo.count() <= 2) {
-			ThreadingTicket ticketThread = new ThreadingTicket(accountId, token, option, flagRepo, lastRepo, dataRepo,
-					intervalRepo);
-			ticketThread.start();
-		} else {
-			System.out.println("===== Still too many rows at DB =====");
-		}
-
-		return new ResponseEntity<Object>(response, HttpStatus.OK);
+		return new ResponseEntity<Object>(response, responseCode);
+	}
+	
+	public Flag newAccountFlag(String accountId) {
+		Flag flagging = flagRepo.save(new Flag(0, accountId, FlagStatus.NEW, 0, 3));
+		return flagging;
 	}
 
 	private void doSaveDataEntryDb(long id, String accountId, String postId, ArrayList<Object> extResource) {
@@ -280,7 +305,6 @@ public class Instagram {
 		/* GET COMMENT ID */
 		String commentId = paramMap.get("parent_id").split("-")[2];
 		String mediaId = paramMap.get("thread_id").split("-")[2];
-		// String mediaUrl = paramMap.get("thread_id").split("-")[4];
 		String igId = paramMap.get("thread_id").split("-")[3];
 		String message = paramMap.get("message").toString();
 		JSONObject metadata = new JSONObject(paramMap.get("metadata").toString());
@@ -288,21 +312,29 @@ public class Instagram {
 		HitApi call = new HitApi();
 		Entity ent = new Entity();
 
+		HttpStatus reponseCode;
+
 		if (metadata.getString("option").equals("1")) {
-			postComment = call.hit(
+			postComment = call.hitAuth(
 					ent.createComment(mediaId, URLEncoder.encode(message, "UTF-8"), metadata.getString("token")),
-					"POST");
+					"POST", errorRepo, igId + " - Channelback");
 		} else {
-			postComment = call.hit(
+			postComment = call.hitAuth(
 					ent.replyComment(commentId, URLEncoder.encode(message, "UTF-8"), metadata.getString("token")),
-					"POST");
+					"POST", errorRepo, igId + " - Channelback");
 		}
 
 		HashMap<String, Object> response = new HashMap<>();
-		response.put("external_id", "cif-comment-" + postComment.getString("id") + "-" + igId /* + "-" + mediaUrl */);
-		response.put("allow_channelback", true);
 
-		return new ResponseEntity<Object>(response, HttpStatus.OK);
+		if (postComment.has("")) {
+			reponseCode = HttpStatus.INTERNAL_SERVER_ERROR;
+		} else {
+			reponseCode = HttpStatus.OK;
+			response.put("external_id", "cif-comment-" + postComment.getString("id") + "-" + igId);
+			response.put("allow_channelback", true);
+		}
+
+		return new ResponseEntity<Object>(response, reponseCode);
 	}
 
 	@RequestMapping("/saveclient")
